+++ CHANGELOG.md
@@ -4,14 +4,14 @@
 
 ### 🎨 Enhancements
 
+- JSON schema enhancements for working with data files
 - [#330](https://github.com/estruyf/vscode-front-matter/issues/330): Allow custom scripts to easily update front matter
 - [#331](https://github.com/estruyf/vscode-front-matter/issues/331): Added functionality to run other type of scripts
+- [#332](https://github.com/estruyf/vscode-front-matter/issues/332): New `dataFile` field which allows you to create data file references
 - [#333](https://github.com/estruyf/vscode-front-matter/issues/333): Automatically mark Jekyll posts in `_drafts` folder as draft
 - [#335](https://github.com/estruyf/vscode-front-matter/issues/335): Merge media snippets with content snippets to allow you to define multiple media snippets and use these in your content
 - [#336](https://github.com/estruyf/vscode-front-matter/issues/336): Support added for inverting the draft field so that SSGs/authors can use a published field instead
 - [#337](https://github.com/estruyf/vscode-front-matter/issues/337): Allow multiple front matter types to be used.
-
-### ⚡️ Optimizations
 
 ### 🐞 Fixes